br/gc: move global functions to manager_global.go and remove api.go
